test(delegation): boundary-fuzz expires_at across monotonic ledger advancement
- Add tests/expiry_boundary.rs: 40 test cases covering 5 boundary offsets
  × 4 ledger sequencing patterns × 2 entry points (delegate,
  execute_delegated_delegate)
- Validate lower bound (expires_at > now) and upper bound
  (expires_at ≤ now + MAX_DELEGATION_DURATION) with strict comparator
  coverage
- Test u64::MAX rejection paired with #372 duration bound
- Verify nonce is NOT consumed on expiry rejection in delegated path
- Monotonic ledger safety: confirm no timestamp capture drift across
  sequential calls, forward jumps, and backward+forward patterns
- Add docs/expiry-boundaries.md with boundary table and security notes
- Fix credence_errors: remove duplicate enum variants (DomainMismatch et al.
  defined in both Bond 218 and Delegation 504 ranges), add missing verifier
  variants (UnknownScheme, VerifierAlreadyRegistered, VerifierNotRegistered,
  VerificationFailed)
- Fix Soroban type compatibility: u8 → u32 in contract public interfaces
  (register_verifier, get_verifier, DataKey::Verifier, DelegatedActionPayload.scheme)

+++ contracts/credence_delegation/src/domain.rs
@@ -80,7 +80,7 @@ pub struct DelegatedActionPayload {
     pub nonce: u64,
     /// The signature scheme: Ed25519, Secp256r1, or MLDSA44.
     /// Defaults to Ed25519 for backwards compatibility with legacy payloads.
-    pub scheme: u8,
+    pub scheme: u32,
 }
 
 /// Validates that the fields in `payload` match the parameters supplied at the
